✨ Feature: Reintroduce Syntax class with highlightBlock and addPattern methods
